resource_mgmt: move storage alert status from storage
This will now be used in cloud_storage's cache class
as well, and it is not desirable to include the whole
set of storage:: types from there.
