+++ src/v/cluster/node/local_monitor.h
@@ -11,6 +11,7 @@
 #pragma once
 #include "cluster/node/types.h"
 #include "config/property.h"
+#include "resource_mgmt/storage.h"
 #include "storage/types.h"
 #include "units.h"
