+++ src/v/resource_mgmt/CMakeLists.txt
@@ -2,6 +2,7 @@ v_cc_library(
   NAME resource_mgmt
   SRCS
     available_memory.cc
+    storage.cc
   DEPS
     Seastar::seastar
   )
