+++ src/v/storage/types.cc
@@ -31,21 +31,6 @@ model::offset stm_manager::max_collectible_offset() {
     return result;
 }
 
-std::ostream& operator<<(std::ostream& o, const disk_space_alert d) {
-    switch (d) {
-    case disk_space_alert::ok:
-        o << "ok";
-        break;
-    case disk_space_alert::low_space:
-        o << "low_space";
-        break;
-    case disk_space_alert::degraded:
-        o << "degraded";
-        break;
-    }
-    return o;
-}
-
 std::ostream& operator<<(std::ostream& o, const disk& d) {
     fmt::print(
       o,
